// src/interfaces/events.ts
export enum SOCKET_EVENTS {
  JOIN_PRESENTATION = 'join_presentation',
  USER_JOINED = 'user_joined',
  USER_LEFT = 'user_left',
  LEAVE_PRESENTATION = 'leave_presentation',
  UPDATE_USER_ROLE = 'update_user_role',
  USER_ROLE_UPDATED = 'user_role_updated',
  ADD_SLIDE = 'add_slide',
  SLIDE_ADDED = 'slide_added',
  REMOVE_SLIDE = 'remove_slide',
  SLIDE_REMOVED = 'slide_removed',
  DISCONNECT = 'disconnect',
  ERROR = 'error',
}